decentralized-identity/requirements.txt
# Requirements for aicache decentralized identity system

# Core dependencies
fastapi==0.85.0
uvicorn==0.18.0
sqlalchemy==1.4.0
aiosqlite==0.17.0
redis==4.3.0
prometheus-client==0.15.0
elasticsearch==8.4.0
pyyaml==6.0
pydantic==1.10.0
cryptography==38.0.0
pyjwt==2.6.0
requests==2.28.0
pandas==1.5.0
numpy==1.23.0
scikit-learn==1.1.0
statsmodels==0.13.0
torch==1.12.0
torchvision==0.13.0
torchaudio==0.12.0
transformers==4.21.0
langchain==0.0.156

# Additional dependencies for decentralized identity
didkit==0.3.0
ssi-sdk==0.1.0
vcx==0.30.0
indy-sdk==1.15.0
ursa==0.3.2